Fix bug when stashing
There is a bug introduced no later than v2.7.4 but fixed as early as
v2.37.1 with `git stash -u`. If there is a folder that is not
git-ignored but contains files which are exclusively git-ignored, then
all the contents of that file will be deleted after running `git stash
-u`, but the contents will not re-emerge after subsequently running `git
stash pop`. In this commit, I make the entire folder `data/` be
git-ignored so as to avoid this bug in the future.